Update for Rust 1.34
This fixes all the warnings and errors reported by rustc 1.34 at build
time. A debug feature is added to the cargo build to make it easier to
toggle debbuging logs at build time.

It was no longer possible to build enigo 0.0.11, bumping the dependency
to 0.0.12 was enough to pass the build and have a working miditran, but
some of the types were deprecated and this introduced new warnings.