Add --mappings option to read a note mapping file
There was already a commented-out partial implementation, so I just
extended it a bit and added a command-line argument to specify the
filename.

This change also includes a sample mappings file for the Alesis Nitro
drumkit using the default keyboard assingments for Clone Hero.
Only the 4 pads are usable at this time, but pro-drums support is
rumored to be coming soon, so I tentatively mapped the cymbals as well.

No setup was required other than hooking up the USB cable and starting
the daemon:

  miditrans -d 'Alesis Nitro 28:0' -f mappings/alesis-nitro.txt

Enjoy!